fix hash impl
