+++ turbopack/crates/turbo-rcstr/src/lib.rs
@@ -291,6 +291,7 @@ impl Hash for RcStr {
             DYNAMIC_TAG => {
                 let l = unsafe { deref_from(self.unsafe_data) };
                 state.write_u64(l.hash);
+                state.write_u8(0xff);
             }
             INLINE_TAG => {
                 self.as_str().hash(state);
@@ -398,8 +399,6 @@ mod napi_impl {
 mod tests {
     use std::mem::ManuallyDrop;
 
-    use rustc_hash::FxHasher;
-
     use super::*;
 
     #[test]
@@ -452,19 +451,5 @@ mod tests {
             }
         };
         assert_eq!(STR, RcStr::from("hello"));
-    }
-
-    #[test]
-    fn test_hash() {
-        let long_string = "A very long long long string that would not be inlined";
-
-        let str = RcStr::from(long_string);
-        assert_eq!(fxhash(str), fxhash(long_string));
-    }
-
-    fn fxhash<T: Hash>(value: T) -> u64 {
-        let mut hasher = FxHasher::default();
-        value.hash(&mut hasher);
-        hasher.finish()
     }
 }
